ch 1 & 4 update version3

+++ chapter 1_principles of Object-Oriented programming/chapter 1_class_code/question_21_code.cpp
@@ -3,7 +3,7 @@
 #include<iomanip>
 using namespace std;
 
-float dg_rd(int x);
+float dg_rd(float x);
 int factorial(int f);
 
 float dg_rd(float x)
